# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_dgrco_quikslate_internal_domain.Shift:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      location_id:
        type: string
      position_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.ShiftStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_dgrco_quikslate_internal_domain.ShiftDetail:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      location_id:
        type: string
      position_id:
        type: string
      position_name:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.ShiftStatus'
      updated_at:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  github_com_dgrco_quikslate_internal_domain.ShiftStatus:
    enum:
    - draft
    - assigned
    - uncovered
    - covered
    - cancelled
    type: string
    x-enum-varnames:
    - Draft
    - Assigned
    - Uncovered
    - Covered
    - Cancelled
  github_com_dgrco_quikslate_internal_response.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.MultipleShiftResponse:
    properties:
      shifts:
        items:
          $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.ShiftDetail'
        type: array
    type: object
...
  internal_handler.SimpleResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_handler.SingleShiftResponse:
    properties:
      shift:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.Shift'
    type: object
...
  internal_handler.assignShiftRequest:
    properties:
      user_id:
        type: string
    type: object
...
  internal_handler.createShiftRequest:
    properties:
      end_time:
        type: string
      position_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.ShiftStatus'
      user_id:
        type: string
    type: object
...
  internal_handler.updateShiftRequest:
    properties:
      end_time:
        type: string
      position_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.ShiftStatus'
    type: object
info:
  contact: {}
  description: Scheduling API for businesses, locations, positions, shifts, and employees.
  title: QuikSlate API
  version: "1.0"
paths:
...
  /businesses/{businessId}/locations/{locationId}/shifts:
    get:
      description: 'List the shifts at a location overlapping [from, to), joined with
        the assignee''s and position''s names. Both range params are required (there
        is no unbounded listing), and the range may not exceed 90 days. Cancelled
        shifts are included. Draft and uncovered shifts are returned only to Admin,
        LocationLead, and Manager: an unpublished plan and an unfilled slot are both
        part of building a schedule rather than reading one, so Employees see neither.
        Any member of the location may call this.'
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Range start, RFC3339 (inclusive)
        in: query
        name: from
        required: true
        type: string
      - description: Range end, RFC3339 (exclusive)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultipleShiftResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: List shifts
      tags:
      - shifts
    post:
      consumes:
      - application/json
      description: 'Create a new shift at a location, optionally pre-assigned to a
        user. Status must be "draft", "assigned" (requires user_id), or "uncovered"
        (requires no user_id). Returns 409 if a pre-assigned user already has an overlapping
        shift. Admin, Manager, or LocationLead. A pre-assigned user must be able to
        see the shift: they need a role at this location, or business-admin status.'
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SingleShiftResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Create shift
      tags:
      - shifts
  /businesses/{businessId}/locations/{locationId}/shifts/{shiftId}:
    delete:
      description: Delete a shift. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift ID
        in: path
        name: shiftId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete shift
      tags:
      - shifts
    get:
      description: Get a single shift by ID. Any member of the shift's location (Admin,
        LocationLead, Manager, or Employee).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift ID
        in: path
        name: shiftId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SingleShiftResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Get shift
      tags:
      - shifts
    patch:
      consumes:
      - application/json
      description: Partially update a shift's status, position_id, start_time, and/or
        end_time. Status cannot be set to "assigned" or "cancelled" here; use the
        dedicated /assign and /cancel endpoints. Moving an assigned shift's times
        returns 409 if it would overlap another of that user's shifts. Admin, Manager,
        or LocationLead; if the shift is assigned, caller must be authorized to act
        on the assigned user's role (see canActOnRole).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift ID
        in: path
        name: shiftId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Update shift
      tags:
      - shifts
  /businesses/{businessId}/locations/{locationId}/shifts/{shiftId}/assign:
    patch:
      consumes:
      - application/json
      description: Assign a shift to a user and set its status to "assigned". Returns
        409 if the user already has an overlapping shift. Admin, Manager, or LocationLead.
        The target must be able to see the shift (they need a role at this location,
        or business-admin status), and the caller must outrank them (see canActOnRole),
        except when assigning to themselves.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift ID
        in: path
        name: shiftId
        required: true
        type: string
      - description: Target user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.assignShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Assign shift
      tags:
      - shifts
  /businesses/{businessId}/locations/{locationId}/shifts/{shiftId}/cancel:
    patch:
      description: Mark a shift as cancelled. Admin, Manager, or LocationLead; if
        the shift is assigned, caller must be authorized to act on the assigned user's
        role (see canActOnRole).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift ID
        in: path
        name: shiftId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Cancel shift
      tags:
      - shifts
  /businesses/{businessId}/locations/{locationId}/shifts/{shiftId}/unassign:
    patch:
      description: Remove the assigned user from a shift. Admin, Manager, or LocationLead;
        caller must be authorized to act on the assigned user's role (see canActOnRole).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Shift ID
        in: path
        name: shiftId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Unassign shift
      tags:
      - shifts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"